refactor: split validator.go into focused modules
- Split 844-line validator.go into 5 focused files:
  - validator.go (210 lines): main validation logic
  - comparisons.go (294 lines): type comparison methods
  - spanner_client.go (188 lines): Spanner connection and queries
  - types.go (45 lines): type definitions and constants
  - utils.go (76 lines): utility functions

- Optimize string concatenation using strings.Builder
- Extract magic numbers to named constants
- Improve code maintainability and readability

No functional changes - all tests pass.
